@@ -1,26 +1,26 @@
 {
-  "_from": "readdirp@~3.4.0",
-  "_id": "readdirp@3.4.0",
+  "_from": "readdirp@~3.5.0",
+  "_id": "readdirp@3.5.0",
   "_inBundle": false,
-  "_integrity": "sha512-0xe001vZBnJEK+uKcj8qOhyAKPzIT+gStxWr3LCB0DwcXR5NZJ3IaC+yGnHCYzB/S7ov3m3EEbZI2zeNvX+hGQ==",
+  "_integrity": "sha512-cMhu7c/8rdhkHXWsY+osBhfSy0JikwpHK/5+imo+LpeasTF8ouErHrlYkwT0++njiyuDvc7OFY5T3ukvZ8qmFQ==",
   "_location": "/readdirp",
   "_phantomChildren": {},
   "_requested": {
     "type": "range",
     "registry": true,
-    "raw": "readdirp@~3.4.0",
+    "raw": "readdirp@~3.5.0",
     "name": "readdirp",
     "escapedName": "readdirp",
-    "rawSpec": "~3.4.0",
+    "rawSpec": "~3.5.0",
     "saveSpec": null,
-    "fetchSpec": "~3.4.0"
+    "fetchSpec": "~3.5.0"
   },
   "_requiredBy": [
     "/chokidar"
   ],
-  "_resolved": "https://registry.npmjs.org/readdirp/-/readdirp-3.4.0.tgz",
-  "_shasum": "9fdccdf9e9155805449221ac645e8303ab5b9ada",
-  "_spec": "readdirp@~3.4.0",
+  "_resolved": "https://registry.npmjs.org/readdirp/-/readdirp-3.5.0.tgz",
+  "_shasum": "9ba74c019b15d365278d2e91bb8c48d7b4d42c9e",
+  "_spec": "readdirp@~3.5.0",
   "_where": "D:\\Fyp\\Server\\fypdb\\node_modules\\chokidar",
   "author": {
     "name": "Thorsten Lorenz",
@@ -48,14 +48,15 @@
   "deprecated": false,
   "description": "Recursive version of fs.readdir with streaming API.",
   "devDependencies": {
-    "@types/node": "^13",
+    "@types/node": "^14",
     "chai": "^4.2",
     "chai-subset": "^1.6",
     "dtslint": "^3.3.0",
-    "eslint": "^6.6.0",
+    "eslint": "^7.0.0",
     "mocha": "^7.1.1",
     "nyc": "^15.0.0",
-    "rimraf": "^3.0.0"
+    "rimraf": "^3.0.0",
+    "typescript": "^4.0.3"
   },
   "engines": {
     "node": ">=8.10.0"
@@ -153,5 +154,5 @@
     "nyc": "nyc",
     "test": "npm run lint && nyc npm run mocha"
   },
-  "version": "3.4.0"
+  "version": "3.5.0"
 }
