@@ -1,8 +1,8 @@
 import {NavigationContainer} from '@react-navigation/native';
 import {createStackNavigator} from '@react-navigation/stack';
 import Axios from 'axios';
-import React, {Component} from 'react';
-
+import React, {Component, useState} from 'react';
+import * as useractions from '../redux/actions/apiuseraction';
 import {
   SafeAreaView,
   StyleSheet,
@@ -18,30 +18,28 @@ import {
   TouchableOpacity,
 } from 'react-native';
 import {ceil} from 'react-native-reanimated';
+import {useDispatch} from 'react-redux';
 
-export default class Login extends Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      username: '',
-      password: '',
-      id: '',
-    };
-  }
-  press = () => {
-    this.props.navigation.navigate('Signup');
+export default function Login({navigation}) {
+  const [username, setUsername] = useState('');
+  const [password, setPassword] = useState('');
+  const [id, setId] = useState('');
+  const dispatch = useDispatch();
+  const press = () => {
+    navigation.navigate('Signup');
   };
-  loginpress = () => {
+  const loginpress = () => {
     let user = {
-      username: this.state.username,
-      password: this.state.password,
+      username: username,
+      password: password,
     };
     Axios.post('http://192.168.10.7:3000/users/login', user)
       .then((res) => {
         if (res.status == 200) {
-          this.id = res.data.User._id;
-          this.props.navigation.navigate('Main');
-          // this.props.navigation.navigate("Main");
+          // setId(res.data.User._id);
+          // dispatch(useractions.ApiloginUser(res.data.User));
+
+          navigation.navigate('Main');
         }
       })
       .catch((error) => {
@@ -50,75 +48,71 @@ export default class Login extends Component {
       });
   };
 
-  render() {
-    return (
-      <View style={styles.container}>
-        <StatusBar hidden></StatusBar>
-        <View style={styles.header}>
-          <Image
-            source={require('../header.jpeg')}
-            style={styles.imageBackground}></Image>
+  return (
+    <View style={styles.container}>
+      <StatusBar hidden></StatusBar>
+      <View style={styles.header}>
+        <Image
+          source={require('../header.jpeg')}
+          style={styles.imageBackground}></Image>
+      </View>
+
+      <View style={styles.footer}>
+        <Text
+          style={[
+            styles.title,
+            {
+              marginTop: 50,
+            },
+          ]}>
+          Name
+        </Text>
+        <View style={styles.action}>
+          <TextInput
+            placeholder="Your name...."
+            style={styles.textInput}
+            // value="a"
+            onChangeText={(text) => {
+              setUsername(text);
+            }}
+          />
         </View>
 
-        <View style={styles.footer}>
-          <Text
-            style={[
-              styles.title,
-              {
-                marginTop: 50,
-              },
-            ]}>
-            Name
-          </Text>
-          <View style={styles.action}>
-            <TextInput
-              placeholder="Your name...."
-              style={styles.textInput}
-              // value="a"
-              onChangeText={(text) => {
-                this.setState({username: text});
-              }}
-              onSubmitEditing={() => this.password.focus()}
-            />
-          </View>
+        <Text
+          style={[
+            styles.title,
+            {
+              marginTop: 20,
+            },
+          ]}>
+          Password
+        </Text>
+        <View style={styles.action}>
+          <TextInput
+            secureTextEntry
+            placeholder="Your password.."
+            style={styles.textInput}
+            // value="a"
+            onChangeText={(text) => {
+              setPassword(text);
+            }}
+          />
+        </View>
 
-          <Text
-            style={[
-              styles.title,
-              {
-                marginTop: 20,
-              },
-            ]}>
-            Password
-          </Text>
-          <View style={styles.action}>
-            <TextInput
-              secureTextEntry
-              placeholder="Your password.."
-              style={styles.textInput}
-              // value="a"
-              onChangeText={(text) => {
-                this.setState({password: text});
-              }}
-              ref={(input) => (this.password = input)}
-            />
-          </View>
-
-          <View style={styles.btn}>
-            <TouchableOpacity onPress={this.loginpress}>
-              <Text style={styles.btntxt}>Login</Text>
-            </TouchableOpacity>
-          </View>
-          <View style={styles.signUp}>
-            <Text style={{color: 'black'}}>New user? </Text>
-            <TouchableOpacity onPress={this.press}>
-              <Text style={{color: 'brown'}}>Signup...</Text>
-            </TouchableOpacity>
-          </View>
+        <View style={styles.btn}>
+          <TouchableOpacity onPress={loginpress}>
+            <Text style={styles.btntxt}>Login</Text>
+          </TouchableOpacity>
+        </View>
+        <View style={styles.signUp}>
+          <Text style={{color: 'black'}}>New user? </Text>
+          <TouchableOpacity onPress={press}>
+            <Text style={{color: 'brown'}}>Signup...</Text>
+          </TouchableOpacity>
         </View>
       </View>
-    );
-  }
+    </View>
+  );
 }
 
 const styles = StyleSheet.create({
@@ -174,6 +168,6 @@ const styles = StyleSheet.create({
   signUp: {
     flexDirection: 'row',
     justifyContent: 'center',
-    marginTop: 120,
+    marginTop: 100,
   },
 });
