@@ -75,7 +75,7 @@ export default class ProfileScreen extends Component {
           <TouchableRipple
             onPress={() => {
               // console.log(this.props.navigation);
-              this.props.navigation.navigate('ChatScreen');
+              this.props.navigation.navigate('Signup');
             }}>
             <View style={styles.menuItem}>
               <Icon name="md-information-circle" color="#dc143c" size={25} />
