@@ -1,13 +1,40 @@
-import React from 'react';
+import React, {useState, useEffect} from 'react';
 import {Dimensions} from 'react-native';
-import {StyleSheet, Text, View, TouchableHighlight} from 'react-native';
+import {
+  StyleSheet,
+  Text,
+  View,
+  TouchableHighlight,
+  ImageBackground,
+} from 'react-native';
 import {Avatar, Accessory} from 'react-native-elements';
 import {TouchableOpacity} from 'react-native-gesture-handler';
+import firestore from '@react-native-firebase/firestore';
+import {useDispatch} from 'react-redux';
+import {chatAction} from '../redux/actions/chataction';
 
-const Single = ({navigation}) => {
+const Single = ({navigation, id, chatName, access, photo}) => {
+  const dispatch = useDispatch();
+  const [info, chatinfo] = useState([]);
+  const db = firestore();
+  useEffect(() => {
+    db.collection('chats')
+      .doc(id)
+      .collection('messages')
+      .orderBy('timestamp', 'desc')
+      .onSnapshot((snapshot) =>
+        chatinfo(snapshot.docs.map((doc) => doc.data())),
+      );
+  }, [id]);
+  useEffect(() => {
+    return () => {
+      console.log('cleaned up');
+    };
+  }, []);
   return (
     <TouchableOpacity
       onPress={() => {
+        dispatch(chatAction(id, chatName, access, photo));
         navigation.navigate('Convo');
       }}>
       <View style={styles.contain}>
@@ -16,15 +43,14 @@ const Single = ({navigation}) => {
             size="medium"
             rounded
             source={{
-              uri:
-                'https://s3.amazonaws.com/uifaces/faces/twitter/ladylexy/128.jpg',
+              uri: photo,
             }}
           />
         </View>
 
         <View style={{paddingLeft: 10, paddingTop: 3}}>
-          <Text>Ali</Text>
-          <Text>testing</Text>
+          <Text style={{color: 'white'}}>{chatName}</Text>
+          <Text style={{color: 'white'}}>{info[0]?.message}</Text>
         </View>
       </View>
     </TouchableOpacity>
@@ -36,9 +62,9 @@ const styles = StyleSheet.create({
     display: 'flex',
     flexDirection: 'row',
     padding: 10,
+    marginTop: 2,
+    borderTopWidth: 1,
 
-    borderTopWidth: 0.5,
-
-    borderColor: 'black',
+    borderColor: 'grey',
   },
 });
