@@ -1,5 +1,5 @@
 import {state, action} from 'react-redux';
-const initialState = {chat: {chatId: 5, chatName: 'usman'}};
+const initialState = {chat: {chatId: null, chatName: null}};
 const chatReducer = (state = initialState, action) => {
   switch (action.type) {
     case 'SET_CHAT':
@@ -8,8 +8,7 @@ const chatReducer = (state = initialState, action) => {
           chatId: action.chatId,
           chatName: action.chatName,
           Access: action.access,
-          Creator: action.creator,
-          members: action.members,
+          photo: action.photo,
         },
       });
 
