@@ -1,49 +1,62 @@
-var createError = require('http-errors');
-var express = require('express');
-var passport = require("passport")
-var path = require('path');
-var cookieParser = require('cookie-parser');
-var logger = require('morgan');
+var createError = require("http-errors");
+var express = require("express");
+var passport = require("passport");
+var path = require("path");
+var cookieParser = require("cookie-parser");
+var logger = require("morgan");
 var cors = require("cors");
-var indexRouter = require('./routes/index');
-var usersRouter = require('./routes/users');
-var mongoose=require('mongoose');
-const connection = mongoose.connect('mongodb://localhost:27017/fypdb', { useNewUrlParser: true, useUnifiedTopology: true });
-connection.then((db) => {
-  console.log("Connected correctly to server");
-}, (err) => { console.log(err); });
+var indexRouter = require("./routes/index");
+var usersRouter = require("./routes/users");
+var adminRouter = require("./routes/admin");
+
+var mongoose = require("mongoose");
+const connection = mongoose.connect("mongodb://localhost:27017/fypdb", {
+  useNewUrlParser: true,
+  useUnifiedTopology: true,
+});
+connection.then(
+  (db) => {
+    console.log("Connected correctly to server");
+  },
+  (err) => {
+    console.log(err);
+  }
+);
 var app = express();
-
+var bodyParser = require("body-parser");
+app.use(bodyParser.json({ limit: "50mb" }));
+app.use(bodyParser.urlencoded({ limit: "50mb", extended: true }));
 // view engine setup
-app.set('views', path.join(__dirname, 'views'));
-app.set('view engine', 'jade');
+app.set("views", path.join(__dirname, "views"));
+app.set("view engine", "jade");
 
-app.use(logger('dev'));
+app.use(logger("dev"));
 app.use(express.json());
 
 app.use(passport.initialize());
 app.use(passport.session());
-app.use(express.urlencoded({ extended: false }));
+// app.use(express.urlencoded({ extended: false }));
 app.use(cookieParser());
-app.use(express.static(path.join(__dirname, 'public')));
+app.use(express.static(path.join(__dirname, "public")));
 app.use(cors());
-app.use('/', indexRouter);
-app.use('/users', usersRouter);
+app.use("/", indexRouter);
+app.use("/users", usersRouter);
+app.use("/admin", adminRouter);
 
 // catch 404 and forward to error handler
-app.use(function(req, res, next) {
+app.use(function (req, res, next) {
   next(createError(404));
 });
 
 // error handler
-app.use(function(err, req, res, next) {
+app.use(function (err, req, res, next) {
   // set locals, only providing error in development
   res.locals.message = err.message;
-  res.locals.error = req.app.get('env') === 'development' ? err : {};
+  res.locals.error = req.app.get("env") === "development" ? err : {};
 
   // render the error page
   res.status(err.status || 500);
-  res.render('error');
+  res.render("error");
 });
 
 module.exports = app;
