@@ -1,20 +1,23 @@
-var mongoose = require('mongoose');
+var mongoose = require("mongoose");
 
 var schema = mongoose.Schema;
 var passportLocalMongoose = require("passport-local-mongoose");
 
 var userSchema = new schema({
-    
-    email: {
-        type: String
-    },
-    height: {
-        type: String
-    },
-    weight: {
-        type: String
-    }
+  email: {
+    type: String,
+  },
+  height: {
+    type: String,
+  },
+  weight: {
+    type: String,
+  },
+  admin: {
+    type: Boolean,
+    default: false,
+  },
 });
 userSchema.plugin(passportLocalMongoose);
 
-module.exports = mongoose.model('User', userSchema);
+module.exports = mongoose.model("User", userSchema);
